Adds new metrics for some apollo specific env variables (#4364)
Adds some metrics for some specific apollo env variables to show if they
are set or not.
Uses the new metrics macros rather than the hash map to exercise more of
the stack.
As a bonus, the thread that was used to continuously update the gauges
is gone.

I had to make some changes to the asserting to allow asserting of all
metrics rather than just specific ones hence all the snapshot changes.
On the plus side they look better now.

Fixes #4290

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: bryn <bryn@apollographql.com>