fix: update telemetry config in helm chart (#4360)
This patch updates the generated router config to avoid a "router
configuration contains deprecated options" warning on router startup.

I tested this change against the in house router's chart in three ways:

* As it currently is, using the deprecated shape. The results are the
same and the router can start up with it.
* Without any telemetry at all. The resulting config no longer contains
`telemetry`.
* After updating telemetry configuration to the new structure. The
resulting config is correct and the router no longer spits out
deprecation warnings.


<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.