Expose Context Id to Rhai scripts (#4374)
We recently added an ID to Context which uniquely identifies the context
for the duration of a request/response lifecycle. This is now accessible
on Request or Response objects from Rhai scripts.

For example:

```rhai
// Map Request for the Supergraph Service
fn supergraph_service(service) {
    const request_callback = Fn("process_request");
    service.map_request(request_callback);
}

// Generate a log for each Supergraph request with the request ID
fn process_request(request) {
    print(`request id : ${request.id}`);
}
```

Note: We have chosen to expose this Context data directly from
Request/Response objects rather than on the Context object to avoid the
possibility of name collisions (with "id") in the context data.

fixes: #4370

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

I tested this manually by modifying various example logging scripts and
logging out the ID for all supported services with non deferred queries
and deferred queries.

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.