Allowlist benign positive secret detections (#4415)
This PR creates updates the secret allowlisting configuration already in
this repo to allowlist more detected "issues" the detected values are
not actually secrets - they are benign positive detections.
Additionally, this PR introduces some config to help prevent this type
of benign-positive detection in the future by allowlisting `testdata`
directories for a number of detection signatures and adding line
comments to some tests so that future commits that interact with those
lines of code do not create new detections

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [X] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

None that I'm aware of? 

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.