chore: a few minor fixes / improvements (https://github.com/apollographql/federation-next/pull/123)
This PR is a grab-bag of a few things I ran into while getting into API
schema validation. I won't be able to finish that before holiday break
so posting the generic changes first.

- Change uses of `directives.iter().any()` to `directives.has()`, which
is a little easier:
[`DirectiveList::has`](https://docs.rs/apollo-compiler/1.0.0-beta.11/apollo_compiler/ast/struct.DirectiveList.html#method.has)
- Return an `Iterator` instead of a `Vec` from `get_types()`, the user
can still decide to `collect()` it into a vec if needed. Also adds a
parallel `get_directive_definitions()`.
- Return `Option<&LinksMetadata>` instead of `&Option<LinksMetadata>`--
an option containing a reference is a much more useful type than a
reference to an option. You basically always want to prefer an
`Option<&T>` if you can.
- c5193a39afa8ddf50e0ea60577cd3bf80f5c4d32 sets source files on
supergraphs when merging subgraphs; this will let diagnostics point to
source code in the future
- Remove the `Result<>` wrapper around `SpecDefinitions`.
`SpecDefinitions::add` now panics if you try to add a wrong version,
which is correct as it's a programmer error. And it's now much easier to
use the `JOIN_VERSIONS` etc. throughout the code.