Aggregate meter provider also delegates to global meter provider (#4460)
The `AggregateMeterProvider` is used by the Router to allow us to have
several metrics subsystems that forward metrics to different places. In
addition it has enhanced testability.

However, some users may be trying to set up otel by themselves outside
of the regular router lifecycle, and they will receive no metrics.

This PR also makes the `AggregateMeterProvider` forward to the
`GlobalMeterProvider`. The setup for this happens at creation time of
`AggregateMeterProvider` and setting the regular otel meter provider
after this happens will have no effect.

This functionality is not reccommeneded for general users, and may
change in breaking ways in the future.

Fixes #4313


<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: bryn <bryn@apollographql.com>