log rate limiting
we want to add more error logging, but worry that we will end up
spamming the  log collectors if the errors happen on every request.
There is also the worry that traces would become hard to debug if events
are rate limited: we could end up with traces that are obviously
wrong, but without the event that could give some context.

This implements a rate limiter for logs, at the formatter level, so that
traces get their own events properly (traces can get sampled
independently)