Context extensions (#4474)
Plugin authors may need to store items in context that are not
serializable. This could be something like an HTTP client, or something
else that cannot be serialized to a Json object.

This PR renames the existing `private_entries` in context to
`extensions` and adds an accessor.
Care should be take to not hold onto mutex guards for too long, however
this is a non-issue because:
* Users should not be performing blocking operations from within non
async blocks. This is not a new issue and we can't prevent them from
doing this.
* MutexGuard is not send, so users won't be able to hold the lock across
async awaits anyway.

Fixes #4445


<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: bryn <bryn@apollographql.com>