add formatting options for trace_id
resolves #4422

Previously, when the `experimental_response_trace_id` was enabled,
the router would append the trace id, if present, to the headers
of the response in hexadecimal format, left-padded with 0s. This
commit allows specifying through the new config option `format`
whether to have that value be hexadecimal (the previous behaviour
and the new default) or decimal.

Below is an example of the new configuration option in action:

```yaml
telemetry:
  apollo:
    client_name_header: name_header
    client_version_header: version_header
  exporters:
    tracing:
      experimental_response_trace_id:
        enabled: true
        header_name: trace_id
        format: decimal # Optional, defaults to hexadecimal
```