do not propagate Accept, Accept-Encoding and Content-Encoding headers (#4535)
Fix #4312
Fix #4398

Potential regression caused by this PR: the only way to activate
subgraph request compression right now is to have the `Content-Encoding`
header set. This header was not part of the reserved headers before, and
it is now, so if a router was configured to propagate that header
(either explicitely or by overly large matching), then now the subgraph
request will not be compressed.
I do not think it is a good idea to have subgraph compression driven by
the headers. That header should instead be set by the subgraph service
directly, from an option set in configuration