Fix valid APOLLO_TELEMETRY_DISABLED flag options
resolves #3023

This commit allows for a more valid range of values for the env var
APOLLO_TELEMETRY_DISABLED, which should fix inconsistencies with
rover's behaviour when disabling telemetry.

There should probably be a way to pass the top-level command args
to other portions of the codebase so that telemetry can be respected
across the codebase using the same deserialization logic.