Progressive override support (#4521)
This change brings support for progressive `@override`, which allows
dynamically overriding root fields and entity fields in the schema. This
feature is enterprise only and requires a license key to be used.

Support for this is implemented via plugin and support for a new context
key whose value is provided as options to query planning.

We expect users to interact with and resolve labels themselves via
coprocessors, so we're providing an interface via a known context key in
order to communicate to and from. The coprocessors need to know which
labels are available in the schema and they need to tell the router
which labels to set.

We also expect users to use a "special-case" label format `percent(x)`
which the plugin will resolve out of the box.

All of this is to say, the plugin has a few responsibilities:
1) At the beginning of a request (`router_service`), the plugin will add
all (non-percent) labels from the schema to the known context key
`apollo_override::unresolved_labels`.
2) Coprocessors (during `supergraph_request`) may read from this context
entry and append labels to be overridden to the
`apollo_override::labels_to_override` context entry.
3) The plugin (during its `supergraph_request`) will resolve
percentage-based labels and append to the
`apollo_override::labels_to_override` context entry.
4) Additionally, the plugin will traverse the operation to obtain only
the relevant labels for the given operation. It will then take the
intersection of those labels and all labels in the
`apollo_override::labels_to_override` context entry to ensure the set of
labels is minimized, sorted, etc.

It's important to note that the set of overridden labels is an input to
query planning, and thus becomes part of the cache key. For every
relevant label in an operation, the number of potential QP cache entries
doubles. Step 4 is important to ensure we don't create any more cache
entries per operation than is necessary.

Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>