Allow access to the parsed document via context
We are nearing 1.0 of apollo-rs, and we need to allow users to access apollo-rs datastructures without having to jump through hoops.

Allow access of the parsed document via context extension, but clearly documenting that was may still make breaking changes.

We will create a way to access schema via a ParsedSchema extension.

Part of #4602