ci: SECOPS-2454 - scan nightly containers for vulns (#4456)
This PR updates CI on this repo to run vulnerability scans of the
nightly container builds. These scans are run using Wiz, the CSPM tool
in-use by Apollo.

There are a couple of items I'd like to discuss prior to getting this
merged:
1. Right now, the scan is configured to use the default container
scanning policy in Wiz which may not be desirable here. This policy
controls when the scan passes/fails. We may want to craft a custom
policy to be more restrictive than our default.
2. We should sort out how/where notifications are sent if the container
vulnerability scan fails. For a series of reasons, its not
straightforward to use the slack-notification that is already in place
if the steps to build the container fail for some reason.

Mentioning these here, but happy to have the actual conversation about
these whereever - in this issue, in slack, or in a call. Just let me
know?

Fixes SECOPS-2454

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [X] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [X] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.