feat(pq): use `accept-encoding: gzip` when downloading PQL chunks
The server that serves chunks support `accept-encoding: gzip` (and in
fact, the chunks are currently stored GZIPped so getting them compressed
is a server-side no-op) but we forgot to enable compressed streaming
when originally writing the manifest polling code.

In order to use reqwest's auto-gzip feature, you have to enable the gzip
feature on the crate, and this also turns it on by default. While it's
likely that this is a beneficial feature in other parts of Router, I did
my best to disable it in all other uses of reqwest in apollo-router that
I could find. (If nothing else, this was necessary to fix the
`it_compress_response_body` test which uses a reqwest::Client to make
sure that Router returns compressed bodies properly.)