test(ci): Pin to 2022.08.1 CircleCI Windows image, rather than "stable" (#4637)
CircleCI updated their `current` (also known as `stable`) label to a new
version that was published around the time that our Windows CI tests
started [mysteriously and consistently
failing](https://app.circleci.com/pipelines/github/apollographql/router/18909/workflows/3aecc54c-2adc-40d9-acd3-1cb6ef7a63b8/jobs/125047)
on February 8th at approximately 15:00 UTC.

Through logs and conversations with support, we understand this to be
true. To mitigate this and get our builds passing again, we are
hardcoding the builds to the version that we now are able to know with
certainty _was_ working, `windows-server-2019-vs2019:2022.08.1`. The
version which appears to be failing is **at the very least**
`windows-server-2019-vs2019:2023.01.1`, but possibly earlier.

This PR changes the unpinned `stable` tag in our configuration to use
the known-good release and we will then release and iterate _up_ from
there, as much as possible.

This same PR will be applied to similar failures in `rover` as
https://github.com/apollographql/rover/pull/1846