Fix active_session_count when future is dropped (#4619)
Fix #4601
We've noticed the `active_session_count` was increasing over time and
never coming back to 0. It seems like this happens because it's possible
for the future to be dropped before we ever get to decrement this
static. Instead, a guard-like struct can be used to decrement on `drop`.

I believe the behavior should be roughly the same (I know there was a
comment on waiting for the response to be fully transmitted before
decrementing), let me know if you have any thoughts.

---------

Co-authored-by: Marc-Andre Giroux <mgiroux@netflix.com>
Co-authored-by: Geoffroy Couprie <apollo@geoffroycouprie.com>