define a trace id for propagated headers when not sampling (#4609)
Fix https://github.com/apollographql/router/issues/4544

When a trace is not sampled (due to the filtering layer we use in the
telemetry plugin), the `OpenTelemetryLayer` is not called, and so a
`TraceId`(the opentelemetry-api one, not the router side one) is never
generated, which means that when reaching the header propagation code,
the current span context is invalid (trace id = 0 and span id = 0) and
so the propagation headers are not sent.
We want the propagation headers to still be sent when the trace is not
sampled, to let the subgraphs know that they should not sample the trace
either.

One option was to explicitely set the propagation headers, but that
depends a lot on which propagation protocol is used. The trick I use
here is generating on the fly a span context with non zero trace id and
span id, so that the headers can be generated.