Prevent subgraph request body being loaded into memory if compression is not enabled Fixes #4648
Previously, even if compression was not enabled the entire body of the request was loaded into memory.

This PR adds logic to detect either `identity` or a missing content-type, and allows the request through untouched.

Tests have been added for content type detection and the act of draining the request.