use a non blocking stdout (#4625)
Fixes #4612 

If the router's output is piped into another process, and that process
does not consume that output, it can entirely lock up the router.

In Rust, stdout access is protected by a lock, to prevent multiple
threads from interleaving their writes. And inside that lock, there is a
LineWriter wrapping a BufWriter to batch line writes.

When the process receiving the output from the router does not consume,
then the BufWriter's writes to the stream start to block, which means
the current thread is blocked while holding the stdout lock. And then
any other thread that might want to log something will end up blocked
too, waiting for that lock to be released.

In this PR we mark stdout as non blocking, which means that writing logs
will fail silently when the buffer is full. This is done in
`executable.rs` to avoid disrupting tests, as any use of println could
panic now.
Documentation is updated to call out that in the cases where the router would have hanged before, if a plugin is using `println!` directly, then that call will panic



Co-authored-by: Jesse Rosenberger <git@jro.cc>