Add a metric for federation version (#4713)
To get a better understanding of how many users are still on fed v1,
this PR adds a metric.

I put it under the `lifecycle` key, is that appropriate?

The version is inferred based on the version of the join spec that is
used, as this is closely tied to the federation version. The federation
minor version is not available in a supergraph as far as I know (it
could only be inferred based on features that are in use, but that's not
reliable if someone doesn't use every feature).

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.